+++ src/transform/scale_mmx64_msvc.c
